requirements.txt
polars==1.39.3
python-binance==1.0.36
py-clob-client==0.34.6
numpy==2.2.3
pandas==2.2.3
dateparser==1.4.0
tzlocal==5.3.1
python-dotenv==1.0.1
